requirements.txt
mlflow
dvc
pyspark
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
